backend/requirements.txt
flask
flask-cors
torch
transformers
opencv-python
numpy
pillow